Made test system test NUMLOCALES/.numlocales tests for CHPL_COMM=none
Previously, the use of a NUMLOCALES/.numlocales file
in the test system would cause a directory/test to
be skipped if CHPL_COMM was set to none.  This
check-in removes that condition, implying that if
a particular does not make sense for CHPL_COMM=none,
a SKIPIF/.skipif file should be used to indicate
that it should be skipped instead.  A test that
has a NUMLOCALES/.numlocales test will be run
for CHPL_COMM=none but without setting the number
of locales.

The rationale for this change is an assumption that
the bulk of future tests will make sense to test
under CHPL_COMM==none as well as other settings -- 
e.g., a parallel implementation of a benchmark as 
opposed to a test that a particular feature works 
within an on clause.

Thus, when creating a NUMLOCALES/.numlocales test,
there are now three main cases that it might fall
into:

  - works equally well on 1 or many locales and
    will generate the right output

  - works on 1 or many locales, but generates
    different output that is worth testing for.
    In this case use a ...comm-none.good file to
    indicate the output for the CHPL_COMM=none
    case and a normal .good file to indicate
    the output for the multilocale case.

  - is not worth testing on 1 locale, in which
    case a SKIPIF/.skipif file should be used
    (e.g., CHPL_COMM == none)

Reorganized existing tests to put them into
one of these three categories.  For many
users' multilocale tests, I pushed tests
into a directory named needMultiLocales/
if it fell into the third category (and added
a SKIPIF file to this directory), and a
oneOrMoreLocales/ directory if it fell into
either of the first two categories.

I have tested this quite a bit before
committing, but am not 100% that I got
everything because the send_patch/receive_patch
scripts don't work so well when files are moved
around.  Spot-checking the results, it looks 
clean, though.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13990 3a8e244f-b0f2-452b-bcba-4c88e055c3ca