refactoring of def/use support code
Removed vectors Symbol::defs and Symbol::uses making the Symbol class
significantly more compact.

Replaced compute_sym_uses functions by new functions called
buildDefUseMaps (a more descriptive name).  These now take as
reference arguments two maps from symbols to SymExprs, one for defs
and one for uses.

Added addDef and addUse functions to add defs or uses to these maps in
order to maintain their correctness (and avoid a recomputation).
Replaced instances of this code by calls to these functions.  That is,
there is code that does this but it was not abstracted into a function
before now.

Added freeDefUseMaps function to free these def and use maps.

Added stylized for-loops for_defs and for_usees to iterate over the
defs or uses of a symbol as given by a map.  For example,

  for_uses(use, useMap, sym) { ...

loops over the uses of symbol 'sym' as specified in the map 'useMap'
such that the SymExpr use is put in macro-declared variable 'use'.

Moved buildDefUseSets to astutil.cpp as it provides similar
functionality to buildDefUseMaps.  This function produces a set of all
of the defs and a set of all of the uses occuring in a given function
for a given list of symbols.  Reimplemented this function so that it
no longer relies on the def and use maps being created.  Refactored
the code to share a common core.

Removed function buildDefsVectorMap which was specific to reaching
definitions and used the old defs and uses mechanism.  This is now
computed more simply in the reaching definitions code.

Optimized reference propagation to build much smaller sets of uses and
defs and to avoid doing so twice.  All of the performance gained in
this checkin comes from this change.  Removing the fields and using
hash tables instead accounted for a minor performance drop though I
still think it is worthwhile moving forward as it cleans up the
representation and I believe it should decrease the maximum amount of
memory that the compiler requires.

This change improves compiler performance slightly:

total compilation time                           before  after
arrays/deitz/jacobi                               6.81    6.75
types/tuple/diten/tupleOfTupleOfTupleOfTuple     17.50   15.44


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13994 3a8e244f-b0f2-452b-bcba-4c88e055c3ca