optimized scalar replacement
Restructured, and largely rewrote, the scalar replacement
optimization.  This localizes the iterative parts and greatly improves
performance:

total compilation time                           before  after
arrays/deitz/jacobi                               6.56    6.44 ( 1.8%)
types/tuple/diten/tupleOfTupleOfTupleOfTuple     15.39   10.55 (31.4%)

For types/tuple/diten/tupleOfTupleOfTupleOfTuple, the time spent in
the scalar replacement pass is cut from 4.88 to 0.62 seconds.

Refactored reference propagation to share code between it and scalar
replacement without making a call to reference propagation on each
iteration.

The generated code is nearly identical.  For the stream benchmark, the
same iterator class is still not scalar replaced and the number of
lines of code for stream.c is unchanged.

Added debugging functionality to the scalar replacement pass to make
it easy to see what classes/records are not replaced.  Hopefully this
will make it easier to improve scalar replacement to handle the one
class that gets by for the stream benchmark.

Also in this checkin, added code to the end of function resolution to
prune the default type constructors for references (which are never
used).  This improves inling performance presumably because calls in
this function no longer need to be handled.

For types/tuple/diten/tupleOfTupleOfTupleOfTuple, the time spent in
the inlining pass is cut from 3.14 to 2.93 seconds.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14003 3a8e244f-b0f2-452b-bcba-4c88e055c3ca