Modified the pthreads runtime again this time to make chpl_process_task_list() execute any tasks (not just the first one) that have not yet begun executing.  This allows the fock-1-latest test to run to completion with maxThreads set to 3, for example.
Basically, the way this works is that chpl_process_task_list() is given a list
of tasks created from a cobegin.  It puts all these tasks, except for the first
one, into the task pool, and then schedules these tasks for execution.  It then
executes the first task (which was not put into the task pool), after which it
goes through the list of tasks it was given to check if they have finished
executing.  If not, it will execute the ones that have not already started
executing.

The basic data structure changes for this involve a field in each task list
entry given to chpl_process_task_list() that is set to true whenever the given
task has finished executing, as well as a field pointing to the task pool entry
for the given task.  Also, each task pool entry has a field that specifies
whether that task has begun executing, as well as a pointer back to the
corresponding task list entry so the completed field can be set to true.

Before a thread takes a task off the task pool to start executing it, it first
checks to see if that task has already started executing.  If so, it removes
that task from the task pool, and goes on to the next task in the task pool.
This is done because there is no mechanism to remove arbitrary tasks from the
task pool; the only task that can be easily removed from the task pool is the
first one.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14010 3a8e244f-b0f2-452b-bcba-4c88e055c3ca