added code to remove 'identity defs' to scalar replacement
Added a function to remove moves like move(x, x), which I've called an
identity def since it defines a variable in terms of itself, for
candidates for scalar replacement because useless defs like this can
inhibit the effectiveness of this optimization.  This gets HPCC STREAM
back on track.  Note that this used to be done in the old
implementation of scalar replacement (before I optimized it in r14003)
but it was not enough.  Coupled with a better order, it is enough.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14015 3a8e244f-b0f2-452b-bcba-4c88e055c3ca