added support for tuples of references for var iterators
This change enables the idiom

  for (a, b) in (A, B) do
    a = b;

given the following declaration

  var A, B: [1..3] int;

Prior to this change, references were lost in tuples requiring the
idiom above to be written as

  for (i, b) in (A.domain, B) do
    A(i) = b;

Capturing references in indices makes the language and implementation
significantly more complex.  These changes took 1.5 days of
concentrated effort (and some serious luck) and I expect the time to
maintain this in working order will be non-trivial.  Therefore, I'd
like us to think about the impact of this change especially as we
discuss parallel iterators.  I'm not opposed to reference indices but
I'd like to be able to justify the added complexity moving forward.

Removed two futures and added two more complicated tests of this
idiom.

As a side note, this change impacts compilation performance
significantly because of the streamlined tuple constructors resulting
in significantly fewer lines of inlined code.

total compilation time                           before  after
types/tuple/diten/tupleOfTupleOfTupleOfTuple     10.57   6.32 (40.2%)

This is on par with the time to compile the jacobi test.

The details of this change follow:

Added pragma "allow ref" that can be applied to generic functions to
indicate that they should be instantiated over references (rather than
dereferencing before the instantiation).  This pragma also disables
culling of references passed to reference arguments for these
functions.  Applied this pragma to a new function called
_build_tuple_always_allow_ref that takes a variable number of
arguments and returns the tuple it forms thus allowing tuples to be
created with fields that are references.  Also applied this pragma to
instantiated tuple constructors.

Changed the code that implements zipper iteration in the tuple module
so that it uses the _build_tuple_always_allow_ref function to create
tuples thus allowing references to be captured during zipper
iteration.

Changed preFold support of the tuple parameter access functions so
that they return a reference via the GET_MEMBER_VALUE primitive for
tuples that have reference fields.

Added pragma "expand tuples with values" that can be applied to
functions that use the (...expression) to expand tuples.  It says that
these expansions should be done with the values in the tuple so as not
to always create references to what is in the tuple.  This is
important if the result is passed to a function that has the "allow
ref" pragma.

Added a "_copy" call to the code that is inserted to determine the
return type of a reduction since without this code, we could end up
trying to instantiate the reduction class over references, which
doesn't work because they don't have the necessary pragmas (nor should
they).  This is necessary because we generate code to internal module
functions that would preserve the references.

Simplified the code generated for instantiated tuple constructors to
avoid using the INIT primitive since we may be dealing with tuples of
references.  This is a simplification because it means the
instantiated tuple type constructors and regular constructors are more
alike.

Disabled insertions of a dereference on a non-var functions returned
expression in normalization if the function is a wrapper since
wrappers are not var functions even if they wrap one.

Made the _pass function for references into a var function.

Added initialization of reference fields when initializing iterator
classes to avoid some use-before-def warnings.  I'm beginning to think
that we should disable these warnings on the backend compilers but
haven't convinced myself of this yet.  There is a definite trade-off
because I could imagine that it could help us to find errors and keep
our generated code clean.  On the other hand, it is a conservative
warning message and we know that it doesn't matter for most of the
iterator code.  Thus the extra initialization adds overhead as well as
complexity.

Simplified the code in homogeneous tuple declarations to avoid a
parameter coercion from int to uint.  At some point, this gave me
trouble though it may not be strictly necessary.

Fixed a seg fault in the html view code I ran into during this change.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14018 3a8e244f-b0f2-452b-bcba-4c88e055c3ca