New semantic future: const-ness of indices
Adding the case we discussed in today's meeting
regarding whether or not a loop's indices are
const.  I've believed this should be the case
for some time, but either believed that they
already were or else never went to the effort to
file the future.  Captured some of our discussion
in the .future.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14021 3a8e244f-b0f2-452b-bcba-4c88e055c3ca