added util/compileTestProgram.pl
This script prints and executes a command to execute a test program
from the current directory with the options specified in the test's
.compopts file or the directory's COMPOPTS file.  It also changes the
relative paths of the extra .chpl files listed in these files.  For
example,

  util/compileTestProgram.pl test/release/examples/hpcc/stream.chpl

prints and executes the command

  chpl test/release/examples/hpcc/probSize.chpl test/release/examples/hpcc/stream.chpl

In addition, the script passes any other arguments along to the
underlying chpl command.

Since I often work in the chapel/ directory I find that I often have
to recreate this command.  Aliasing this script to tchpl allows
debugging the compilation of an associative array test with many
compiler options easy.  Write

  tchpl test/associative/deitz/arrays/test_indefinite1.chpl --gdb

to print and execute

  chpl test/associative/deitz/arrays/../../common/printHelp.chpl --main-module=test_indefinite1 test/associative/deitz/arrays/test_indefinite1.chpl


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14023 3a8e244f-b0f2-452b-bcba-4c88e055c3ca