added gNilRef and dtNilRef
This primitive symbol and type is a reference version of nil that can
be used to initialize references.  Moving gNil into a reference
results in an assignment to where the reference refers.  Since gNilRef
has the "ref" pragma, this acts as a move from one reference to
another.  Used gNilRef to initialize iterator classes which can now
contain references in tuples.  Made the insertWideReferences pass
insert temporaries for gNilRef similar to gNil.

This checkin should fix the new gasnet/intel regressions, hopefully.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14025 3a8e244f-b0f2-452b-bcba-4c88e055c3ca