extended ref indices to work with tensor iteration
Added a test case.

Added a future showing that the setter argument used in promoted var
functions that are iterated over is incorrect.  This essentially
requires support for setter in iterators, too.  The future states:

  "The setter argument is not correctly set for var functions that are
   promoted and then iterated over in a zipper context.  This seems to
   be because cullOverReferences does not cull the reference field in
   tuples.  To do this, we'll need to clone the tuple.

   My proposed semantics are that the promotion of a var function that
   is iterated over, should set the setter field to true if the index
   may be written.  In var functions, since references cannot be
   captured, the setter=true version is used if the index is written.
   The distinction has to be more conservative in the iterator case."

Also, note that compiling with baseline breaks everything since the
iterator class stores a reference.  I think the solution is to lower
iterators after references are culled.

Changed a semantic future to a bug as a result since it seems like we
now need var iterators.  The setter in var iterators will have to be
conservatively set to true if the index may be written.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14028 3a8e244f-b0f2-452b-bcba-4c88e055c3ca