Reproduce the ChapelBase change from rev. 12834
The Intel compiler warns about some comparisons between uints and literal 0.
Add an additional check for the param int being equal to zero to avoid doing
the comparison at runtime for that case. Revision 12834 only added these cases
when the param int was on the right hand side of the comparison, add the left
hand side cases as well.

Add these special cases to the .chpl util code used to generate these
functions.

This should fix the recent Intel regressions. Thanks Brad for pointing this
out to me.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14029 3a8e244f-b0f2-452b-bcba-4c88e055c3ca