moved run of prune pass before optimizations
This is an improvement made with Samuel and David during a code overview of the
passes in the compiler.  This makes a substantial performance improvement for
compiling a test that declares "128*int" or another large homogeneous tuple.
(33% compile-time speedup)

Moved the run of the prune pass before the optimizations.  This improves the
performance of inlining (and other optimizations) because we do not inline, for
example, any functions within dead functions (since they are pruned).

Simplified the prune pass and in the process made it so that it could be run
multiple times.  It is not, however.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14031 3a8e244f-b0f2-452b-bcba-4c88e055c3ca