detached visible function data structure from symbol table
Created a new data structure for storing visible functions within the
function resolution pass and removed the visible functions from the
symbol table.

While making this change, I noticed that we used to get visible
functions when resolving a call that were not actually visible.  This
change tightens the semantics though it is unclear whether codes could
have been affected.  None of our tests showed an issue.

Removed special handling of functions like _if_fn to handle
conditional expressions in normalization.

Removed special handling of functions without parentheses in
normalization.  They used to be marked as not visible, but this had no
effect since the function was already in the AST.

Reworked how "main" is found in buildDefaultFunctions since it used to
rely on the maintained visible functions in the symbol table.

The primary reason for the compile-time speedup in this checkin is due
to a slightly streamlined getVisibleFunctions implementation that uses
blocks instead of scopes.  That said, getVisibleFunctions remains the
major bottleneck in function resolution especially when compiling a
test case involving the type 128*int.  I have some simple optimization
ideas to try soon.  The speedup on this test is 9.9%.  On
test/arrays/jacobi, I'm seeing a 2% speedup and the first compilation
time of this test below 6 seconds.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14035 3a8e244f-b0f2-452b-bcba-4c88e055c3ca