resolved passing futures (removed 3 of 4)
The test multilocale/bradc/assignRemoteConst fails with --no-local so
I added this in a .compopts file so that this test fails on all
platforms.  This is necessary now that multilocale tests are run
everywhere with numLocales=1.

The test functions/deitz/iterators/test_promote_inout works because
promotion is handled with zipper iteration and the indices are now
references.

The test extern/bradc/defaultArgsNoLocal now seems to work probably
due to orthogonal improvements to the wrappers.

The test domains/bradc/arrayLinkage works per our semantic futures
discussion.  There is mention of documenting this semantic but I'm not
sure what is meant to be documented.  Reading the spec about ".domain"
it says it returns the domain of the array.  It is a const method, but
the domain is not a copy.  Moreover, the new array's domain is
constant because accessing it via .domain is constant.  Brad, let me
know if you want me to change the spec.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14036 3a8e244f-b0f2-452b-bcba-4c88e055c3ca