refactored and optimized getVisibleFunctions
Refactored getVisibleFunctions by simplifying the code to determine
the next block that needs to be searched.

Optimized getVisibleFunctions by introducing a map from source blocks
that do not need to be searched to the first block that needs to be
searched when searching from the source block.

Compile-time speedup for the 128*int test case: 22%
  Function resolution speedup: 39%

Profiling on this test case shows that getVisibleFunctions is no
longer the primary bottleneck.

Added the 128*int test case.

I'm not seeing a significant speedup on the arrays/deitz/jacobi.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14039 3a8e244f-b0f2-452b-bcba-4c88e055c3ca