made it so that all symbols are defined in the AST
This change makes it so that all symbols and types are defined in the
AST except for the new module symbol 'rootModule' which is the root of
everything and replaces rootScope.  The module 'theProgram' which
defines all other modules and acts as a place to define things in the
compiler is defined within rootModule.  The difference between these
modules is that things in rootModule are not codegened whereas things
in theProgram are codegened.  Also, theProgram is 'in' rootModule.

Along with making the symbol table less integral, this change
simplifies some code (such as determining whether symbols are live)
and uncovers some bugs in the handling of complex types when they are
replaced by records.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14045 3a8e244f-b0f2-452b-bcba-4c88e055c3ca