removed runtime.h/runtime.cpp + fixed intel build + misc.
Moved 'theProgram' into symbol.cpp along with baseModule.  This is
where rootModule is.

Renamed array/file/writer types that refer to those types in the
standard modules and moved these to type.h.

Removed ill-named runtime.h and runtime.cpp which just contained
external things in the standard modules (not the runtime).  Removed
some of these because they were not used.

Reenabled instantiation limit check on the tuple module now that the
tuple 'this' method is special-cased in function resolution.

Removed unnecessary 'return' in getVisibilityBlock, thus fixing the
intel build.

Removed scope lookups in buildDefaultFunctions.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14047 3a8e244f-b0f2-452b-bcba-4c88e055c3ca