Fix an accidental int->bool conversion
In heapAllocateGlobals, the localeID was assigned to a temp bool then compared
with zero with that result put into the same temp. In this case, it worked as
expected since 0 is false, but it is dangerous. Added a second temp that
takes its type from the localeID primitive, and use that in the comparison
instead of the bool.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14048 3a8e244f-b0f2-452b-bcba-4c88e055c3ca