Adding another fragmented stream code
This is a second fragmented stream code that moves the
timers outside the coforall as Steve suggested was
more appropriate.  I'm realizing that the reason I
didn't do this originally is that it required
re-initializing and -verifying the computation on every
iteration with my current local-variable-based
approach.  This makes the test untenable for performance
studies -- it times out.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14056 3a8e244f-b0f2-452b-bcba-4c88e055c3ca