Next version of manually-blocked stream: no refs to Locales/LocaleSpace
Removed all hard-coded references to Locales and LocaleSpace
from the Brads1DBlock distribution, using a target locales
member variable in the distribution instead.  Also added
documentation in the form of comments for most members and
methods in my distribution classes to help with reading the
code.

This raised some pretty interesting issues (which I'll describe
below).  I also think it's an amazing demonstration of the 
power of Chapel's generics and type inference.  Which is to say, 
I think writing the equivalent C++ classes would send anyone into 
fits -- we should try it sometime and see.

In particular, I created a "targetLocs" member in the distribution
class that I just introduced and initialized it to be "Locales"
by default.  I then referred to this array of locales (and an
associative domain of locales) in the other classes by following
their .dom and .dist pointers backwards and things just fell out
very cleanly and neatly.  When I did have error messages, they
were very clear and helpful.

Feature request: I wanted to be able to assert that it was a rank-1
array of locales (where I might even be open to associative/opaque/
enumerated sets of locales being provided) and potentially drop the
default value, but couldn't figure out a clean way to do this, so
simply used the type inference to get the arithmetic case at least.
This relates to our semantic future discussion, wanting syntax for
talking generally about array types or constraints.

One of the (anticipated) interesting challenges I ran into is that
when I hard-code a distribution to be across Locales, I often make
assumptions about the locale IDs in the code which would not work
if the distribution targeted other sets of locales.  Here, I took
the approach of storing the arrays of local domain and array classes
an associative array over the locale type so that it's easy to get
from a locale to its local domain/array class.  The upside of this
approach is that I never had to reason about the logical indices
from the domain of the target set of locales nor about the locale
IDs themselves (which would arguably be a bad idea since they don't
hold any context in a distribution, IMO).  The downside is that
this approach makes it harder for the target locale set to
contain duplicate locale entries -- though perhaps that should be
done by engineering the local domain/array classes to support
multiple blocks/chunks (i.e., one local class per locale which is
set up to understand oversubscription -- this might also help
avoid overheads associated with a blinder oversubscription).
Arguably, the associative array is more expensive to use than a
simpler array as well.  Interestingly, I think one would want
to query a member array's domain, which I don't think we support
yet.  I.e.:  class C { var targetLocs: [?targetLocDom] locale; }
With this, then one could declare the array of local domain/array
classes over targetLocDom.  I guess one could do this with
targetLocs.domain as well.  I left a TODO for myself to give this
a shot and see how it works.

In either case, I think one needs to come up with a total ordering
of the locales (an ID relative to targetLocs.domain essentially)
in order to do the initial mapping of indices to locales.  I did
that here using (sequential) zippered iteration against the range
"0.." which works well enough and only occurs at domain setup
time.  It might be nice to see whether there's something more
elegant that could be done here without loss of generality.

This change cleaned up several iterations that I already had
in which I was iterating over LocaleSpace and then using that
ID to get a locale -- in the new code I can iterate over
targetLocs typically which gives me what I want in most contexts --
a locale.  It could be that if I wasn't using the associative
domain/array approach that this would not work as well, which
would take me back closer to my original approach and would
still be reasonable -- maybe a zippered iteration over the
targetLocs and targetLocs.domain is what would be cleanest
while efficient.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14067 3a8e244f-b0f2-452b-bcba-4c88e055c3ca