fixed function visibility bugs
This checkin removes code that made the block surrounding a generic
function "use" the modules that define any types that instantiated it.
I added a comment asking why we were doing this some time ago and I
think we did this before we started working with "point of
instantiation" semantics.

In removing this code, we no longer pollute the visible function
blocks with non-visible functions, but I had to make a number of fixes
to how we handle points of instantiation.

Fixed a bug in which the type constructor for references would have an
instantiation point in main.  It now is whereever the base type is
defined or, if it is generic, instantiated.

Fixed a bug in which the type constructor could have its instantiation
point in the module where the generic type is defined rather than
where the type is specified.

Fixed a bug in which the value version of an instantiated var function
would not have an instantiation point.

Fixed a bug in which a wrapper would not have an instantiation point.
This is especially important for default wrappers where the types may
be used to declare new variables within the body of the wrapper.

Defined the semantics of functions that are instantiated only via
dynamic dispatch to set the instantiation point to the point where the
type of 'this' is defined or, if it is generic, instantiated.  I'm
writing a test, but it seems to need more work.  Since this is working
on existing tests, I plan to checkin.

Moved all of the _cast functions built for enumerated types in
buildDefaultFunctions in to the baseModule because they need to be
globally visible since they can be inserted into coercion wrappers
(that may not even be generic).

Changed the name of the argument of the writeThis method to 'f' in the
test modules/diten/returnClassDiffModule5 so that it continues to
work.  I captured the issue of dynamic dispatch and named arguments in
a separate more stand-alone test.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14070 3a8e244f-b0f2-452b-bcba-4c88e055c3ca