Next version of hand-written distribution
In previous versions, I had a module-level type "indexType"
that I was using to represent a domain's index type in a
way that didn't require me to mess with generic classes more
than necessary.  In this version, I introduced generic types
into the distribution, domain, and array classes to represent
the index types of the global and local domain representations.
Currently these are both hard-coded to be the same, and
the code would probably need to change to make them different
(in particular, casts would probably need to be used).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14074 3a8e244f-b0f2-452b-bcba-4c88e055c3ca