Generic error message line # future
In doing my distributions prototyping, I came across this
interesting error -- when compiling in non-developer mode,
line numbers for generics of generics of generics are not
propagated through multiple layers very well -- in a sense,
the same sort of cutoff that we do between user code and
compiler code seems to kick in and refer to the outermost
in the nested generics rather than the more accurate
location of the error.

The .good file here is an approximation of what I'd expect.
The main thing to get right is line number 14 rather than 17
-- the details can vary as long as it is clear and
appropriate.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14075 3a8e244f-b0f2-452b-bcba-4c88e055c3ca