Made coforall behave similarly to cobegin in that all the tasks created from the iterations are put on a task list, and then they are all put into the task pool at once.  There is actually no change in behavior except in the pthread version, in which the thread executing the coforall executes the first (and perhaps only) task in the task list, and then puts any remaining tasks into the task pool.  It then goes through the task list and executes any tasks that have not yet begun executing.
The change in chpl_cfg_threads.c is to remove the assumption that there are
at least two tasks in the task list, which is the case for cobegins.  The
change in normalize.cpp ensures that local variables used in the coforall
are allocated on the heap, though the only variable that needs to be on the
heap is the loop variable.  (This will be optimized at a later time.)
Finally, a .execopts file was added to restrict maxThreads to 99 for one of
the tests so no warning is emitted about not being able to create more
threads with maxThreads being unbounded.  This test does not need a high
value for maxThreads.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14082 3a8e244f-b0f2-452b-bcba-4c88e055c3ca