fixed a bug relating to fields of range type
This checkin fixes an interesting generic type case that applies to
all generic types that have default values for every generic field
(including the range type).  The issue is that when written, they are
not generic, yet they were treated as such.  This made it impossible
to declare a field of such a type while using the defaults.  In order
to do this, one must query a field of such a type.

Added bool ArgSymbol::markedGeneric to mark a formal argument as
definitely being generic so that we can distinguish range from
range(?x), for example.  This is also set when building some default
functions where we want the formal argument type to be generic.

Added bool Type::hasGenericDefaults to mark types that are generic but
which have default values for every generic field.  These need to be
handled specially so that an argument of such a type is not treated as
being generic.

Inserted queries of fields in formal range type specifications in the
standard modules so as to mark them as generic ranges.

Updated the specification with a note about how to deal with generic
types which have default values for all of their generic fields.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14084 3a8e244f-b0f2-452b-bcba-4c88e055c3ca