added support for eliding the identifier in a query expression
This allows the identifier to be omitted in after the ? in the
definition of a generic function as in

  def foo(r: range(?)) { }

This is important because otherwise there is no easy way to specify
that the generic type is of interest.

Updated the spec to allow eliding the identifier after ?'s.  Note that
this allows allows this omission anywhere so the following are
equivalent:

  def foo(x) { }
  def foo(x: ?t) { }
  def foo(x: ?) { }

Added a test case.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14086 3a8e244f-b0f2-452b-bcba-4c88e055c3ca