Runtime restructuring for launch utility
This is a simple restructuring of the runtime to make
it more amenable to using a launch utility that shares
code.  I some simple, silly launch utility code in
my local directory, but will save that for a separate
checkin once it's a little further along.  This
change...

* adds an #ifndef LAUNCHER case to chplthreads.h and
  chplcomm.h so that runtime/src/*.c files used by
  a launcher (e.g., chplmem.c) thatinclude these will 
  not require the launcher to link in a full thread
  and comm version of the runtime (since it won't
  use threads or comm).  In the #ifdef LAUNCHER case,
  a few key routines and types are stubbed out to
  make references to them in re-used files innocuous.

* Split the chplcomm.h and chpl_cfg_comm.c files so
  that the routines to specify a default number of
  locales and to verify the number of locales are
  in their own files for direct inclusion in either
  the launcher or the main binary (again, to avoid
  linking in an entire threads or comm version of
  the library).

Next step is to check in an initial launcher that
simply calls a perl script that execs its arguments
as a stand-in for an aprun, mpirun, etc.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14096 3a8e244f-b0f2-452b-bcba-4c88e055c3ca