Add tests defining methods outside of the class
The .future causes a segfault in scope resolution where it should cause an
undefined symbol error.

test_nested_method defines a method as a nested function within main. It is
able to access all of the variables one might expect it to be able to.

nested_class_return_inner defines a class nest where the outer has a method
to return an instance of the inner. Then, the inner instance is used to modify
fields in the outer instance.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14107 3a8e244f-b0f2-452b-bcba-4c88e055c3ca