Another instantiation test
Adding a variation on Steve's 
test_point_of_instantiation2-blc.chpl test in the
sibling directory to this one that has been floating
around for awhile.  This test wondered what
instantiating a function in one module would do
when the same instantitation and/or another was
called in the other module, where the instantiated
function calls a helper function that exists in both
modules.  I didn't know what to expect from this,
but nobody seemed surprised, so let's call it correct.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14110 3a8e244f-b0f2-452b-bcba-4c88e055c3ca