reimplemented the symboltable
This commit message is in parts.

  Readjusted cleanup/scope resolution division of work
  Created a new symboltable
  Removed the old symboltable
  Improved visibility disambiguation in function resolution
  Miscellaneous

Readjusted cleanup/scope resolution division of work
----------------------------------------------------

Moved cleanup pass code to create the class/record hierarchy, process
the 'use' statements, and build the constructors into the scope
resolution pass.  Creating the hierarchy and processing the 'use'
statements requires a symboltable for module and type symbol
resolution.  Building the constructors is best done after creating the
hierarchy because the constructor needs to refer to items in the
parent classes/records.  It is also best done before finishing scope
resolution so that scope resolution can be used to insert the implicit
'this' and 'outer' references.

Created a new symboltable
-------------------------

The new symboltable is a hash table from BaseAST pointers that
correspond to a scope to table entries which are in turn hash tables
from canonicalized strings to symbols.  There is only one symbol
associated with each string.  For overloaded functions, only one
function is stored.

The function newLookup (to be renamed in another checkin) replaces
SymScope::lookup and copies most of the code.

Removed a hack that avoided modules in a first lookup and only looked
again if nothing was found.  This made it impossible to shadow
something by a module.  Inserted module uses of ChapelStandard into
every module to make this unnecessary since we don't want a module's
name, e.g., opaque in release/examples/opaque, to shadow the opaque
type in the rootModule.

Removed code to check for duplicate symbol definitions in the same
scope.  This is now done when creating the symboltable.  As a result,
compiler temporaries are not added to the new symboltable.

Removed the old symboltable
---------------------------

Removed SymScope class and all associated fields in the BaseAST
subclasses as well as the related fields BaseAST::parentScope, and
Symbol::overloadPrev and overloadNext.

Greatly simplified the code to insert and remove AST since this no
longer has to update the parentScope pointers, define symbols in the
symbol table, and create new SymScope class instances.

Removed symboltable printing functions in view.h and view.cpp.

Since the symboltable is not maintained throughout compilation,
removed some crufty code that would remove a function before renaming
it.  The function can now be renamed without removing and reinserting
it.

Improved visibility disambiguation in function resolution
---------------------------------------------------------

Improved function resolution's visibility disambiguation code so that
a function reached via a 'use' is shadowed by those before that 'use'.
I'm not sure this code is working well as mentioned in a comment; it
should be rewritten to possibly share code with scope resolution.
This became necessary now that the standard modules are accessed via
'use' statements rather than by following parentExpr/parentSymbol
pointers.  Removed localFnShadowsUse.future.

Miscellaneous
-------------

Fixed a bug in which creating a method on a non-type would result in a
compiler seg fault.  Removed test_scope_resolve_classes.future.

Performance
-----------

I believe this change results in about a 10% improvement in both
compilation speed and compilation memory usage.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@14116 3a8e244f-b0f2-452b-bcba-4c88e055c3ca